requirements.txt
numpy
matplotlib
python-mnist